embark on the long journey of cleaning up this code
